some flang -> f18 renaming
