+++ include/f18/Config/config.h.cmake
@@ -1,12 +1,12 @@
 /* This generated file is for internal use. Do not include it from headers. */
 
-#ifdef FLANG_CONFIG_H
+#ifdef F18_CONFIG_H
 #error config.h can only be included once
 #else
-#define FLANG_CONFIG_H
+#define F18_CONFIG_H
 
 
-#define FLANG_VERSION            "${FLANG_VERSION}"
+#define F18_VERSION            "${F18_VERSION}"
 
 #endif
