+++ lib/runtime/CMakeLists.txt
@@ -15,5 +15,4 @@
 add_library(FortranRuntime
   ISO_Fortran_binding.cc
   descriptor.cc
-  type.cc
 )
