+++ tools/f18/CMakeLists.txt
@@ -19,5 +19,5 @@ add_executable(f18
 
 target_link_libraries(f18
   FortranParser
-  FlangSemantics
+  F18Semantics
 )
